requirements.txt
python-telegram-bot==22.5
python-decouple==3.8
pillow==12.0.0
filetype==1.2.0